requirements.txt
alabaster==0.7.12
appdirs==1.4.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asn1crypto==0.24.0
astroid==2.11.2
async-generator==1.10
attrs==21.4.0
Automat==20.2.0
awscli==1.22.83
Babel==2.9.1
backcall==0.2.0
bleach==4.1.0
blinker==1.4
bokeh==2.3.3
boto==2.49.0
boto3==1.21.28
botocore==1.24.28
cached-property==1.5.2
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.4
cloud-init==22.1
cmake==3.22.3
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
coremltools==5.2.0
cpplint==1.6.0
cryptography==36.0.2
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
dill==0.3.4
distro-info===0.18ubuntu0.18.04.1
docutils==0.15.2
dparse==0.5.1
ec2-hibinit-agent==1.0.0
entrypoints==0.4
environment-kernels==1.1.1
future==0.18.2
graphviz==0.19.1
h5py==3.1.0
hibagent==1.0.1
httplib2==0.20.4
hyperlink==21.0.0
idna==3.3
imageio==2.15.0
imagesize==1.3.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
incremental==21.3.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.7.0
isort==5.10.1
jedi==0.17.2
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
jsonpatch==1.16
jsonpointer==1.10
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
kiwisolver==1.3.1
language-selector==0.1
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
matplotlib==3.3.4
mccabe==0.7.0
mistune==0.8.4
mpmath==1.2.1
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.4
netifaces==0.10.4
networkx==2.5.1
nose==1.3.7
notebook==6.4.10
numpy==1.19.5
nvidia-ml-py==11.450.51
oauthlib==2.0.6
opencv-python==4.5.5.64
packaging==21.3
PAM==0.4.2
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.19.4
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycparser==2.21
pycrypto==2.6.1
pydot==1.4.2
pygal==3.0.0
Pygments==2.11.2
PyGObject==3.26.1
PyJWT==1.5.3
pylint==2.13.2
pyOpenSSL==17.5.0
pyparsing==3.0.7
pyrsistent==0.18.0
pyserial==3.4
pystache==0.6.0
python-apt==1.6.5+ubuntu0.7
python-dateutil==2.8.2
python-debian==0.1.32
pytz==2022.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
requests==2.27.1
requests-unixsocket==0.1.5
rsa==4.7.2
s3transfer==0.5.2
safety==1.10.3
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
Send2Trash==1.8.0
service-identity==16.0.0
six==1.16.0
snowballstemmer==2.2.0
sos==4.3
Sphinx==4.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
ssh-import-id==5.7
sympy==1.9
systemd-python==234
terminado==0.12.1
testpath==0.6.0
threadpoolctl==3.1.0
tifffile==2020.9.3
toml==0.10.2
tomli==1.2.3
tornado==6.1
tqdm==4.63.1
traitlets==4.3.3
Twisted==22.2.0
typed-ast==1.5.2
typing_extensions==4.1.1
ubuntu-advantage-tools==27.6
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.9
virtualenv==15.1.0
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.0
wrapt==1.14.0
zipp==3.6.0
zope.interface==5.4.0
